rename nav pose tool to goal pose tool
