+++ rviz_common/help/help.html
@@ -48,7 +48,7 @@ p { margin-left: 10px }
   </tr>
   <tr class="even">
     <td class="key">G</td>
-    <td>Switch to the 2D Nav Goal tool.</td>
+    <td>Switch to the 2D Goal Pose tool.</td>
   </tr>
   <tr class="even">
     <td class="key">P</td>
